Go: Move `go list` functions to `toolchain` package
